# public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Dastyare Social — CS",
    "version": "1.0.0",
    "description": "REST API for Dastyare Social — Creator Studio. Manage posts (threads, shorts, media) and stories. Supports JSON and multipart uploads, reactions, views, and likes. Authentication via Your API key. See /llms.txt for LLM-oriented documentation."
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "MediaUploadResponse": {},
...
      "PresignResult": {
        "type": "object",
        "properties": {
          "uploadUrl": {
            "type": "string"
          },
          "key": {
            "type": "string"
          },
          "kind": {
            "type": "string",
            "enum": [
              "image",
              "video",
              "audio",
              "file"
            ]
          },
          "mimeType": {
            "type": "string"
          }
        },
        "required": [
          "uploadUrl",
          "key",
          "kind",
          "mimeType"
        ]
      },
...
    },
...
  },
  "paths": {
    "/upload": {
      "post": {
        "operationId": "post-upload",
        "summary": "Upload Media — Single File",
        "description": "Single-file upload endpoint (multipart/form-data with a `file` field). Server-side upload — buffers the file then uploads to S3. Returns media metadata. For real upload progress, use `/api/upload/presign` + `/api/upload/confirm` (direct browser-to-S3) instead.",
        "tags": [
          "Media"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "x-jsonschema-contentMediaType": "application/octet-stream"
                  }
                },
                "required": [
                  "file"
                ],
                "type": "object"
              }
            }
          },
          "description": "Multipart form data containing a file upload.",
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaUploadResponse"
                },
                "examples": {
                  "example": {
                    "value": {
                      "url": "https://cdn.example.com/media/image/abc.jpg",
                      "key": "media/image/abc.jpg",
                      "kind": "image",
                      "mimeType": "image/jpeg",
                      "size": 12345,
                      "width": 1080,
                      "height": 1920,
                      "duration": 0,
                      "filename": "photo.jpg"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/upload/confirm": {
      "post": {
        "operationId": "post-upload-confirm",
        "summary": "Upload Media — Confirm Direct Upload",
        "description": "After the browser uploads directly to S3 via presigned URL, call this with { key, mimeType, filename, size, width, height, duration } to get the public URL and full media object.",
        "tags": [
          "Media"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "url": {
                      "type": "string"
                    },
                    "key": {
                      "type": "string"
                    },
                    "kind": {
                      "type": "string"
                    },
                    "mimeType": {
                      "type": "string"
                    },
                    "size": {
                      "type": "number"
                    },
                    "width": {
                      "type": "number"
                    },
                    "height": {
                      "type": "number"
                    },
                    "duration": {
                      "type": "number"
                    },
                    "filename": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "url",
                    "key",
                    "kind",
                    "mimeType",
                    "size",
                    "width",
                    "height",
                    "duration",
                    "filename"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "403": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "500": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/upload/presign": {
      "post": {
        "operationId": "post-upload-presign",
        "summary": "Upload Media — Presign URL",
        "description": "Returns a presigned URL the browser can PUT to directly, giving real upload progress. Send { filename, mimeType } in the JSON body.",
        "tags": [
          "Media"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PresignResult"
                }
              }
            }
          },
          "400": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "403": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "500": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  },
...
  "servers": [
    {
      "url": "/api",
      "description": "API server"
    }
  ]
}